Core: Fix a potential race condition with file data sources
`FileDataSource` objects have a list of allowed extensions that is
validated on creation. Up to now, this list was processed in an
unspecified order, which might result in the wrong extension being
picked should one be a suffix of another (this is not the case for any
of our data sources). Instead of processing the extensions in
parallel, we now use a sequential stream, which avoids this problem,
and, as a side effect, fixes #198.