Move local VLog installation out ouf rulewerk-dependencies
By not making the local VLog a dependency of rulewerk-dependency and
installing it only for rulewerk proper, we automatically pick up any
changes in VLog for the build, without any requirements on changing
the rulewerk-dependencies output hash. Since a cached
rulewerk-dependencies prevents a rebuild, we would not notice the
changed rulewerk version, and not produce an updated m2 repository.